Transforms rust binds from exe to crate/lib
