+++ Cargo.toml
@@ -1,0 +1,3 @@
+[workspace]
+members = ["engine-sys", "orchestrator"]
+resolver = "2"
