+++ engine-sys/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "engine-sys"
+version = "0.1.0"
+edition = "2024"
+build = "build.rs"
+links = "engine"
+
+[dependencies]
